+++ src/components/Footer.jsx
@@ -1,7 +1,10 @@
 import React from 'react';
-// styles
 import styles from '../styles/Footer.module.css';
 
+/**
+ * Footer component displaying the site's copyright text.
+ * @returns {JSX.Element} React element representing the Footer component.
+ */
 const Footer = () => {
   return (
     <footer className={styles.footer}>