+++ src/main.jsx
@@ -1,12 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import { Provider } from 'react-redux';
 // routes
 import Rooter from './router';
+// store redux
+import store from './store';
 // styles
 import './styles/index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <Rooter />
+    <Provider store={store}>
+      <Rooter />
+    </Provider>
   </React.StrictMode>
 );